knuth-bendix: change gilman graph node type to uint32_t

+++ include/libsemigroups/knuth-bendix.hpp
@@ -205,7 +205,7 @@ namespace libsemigroups {
 
     Rewriter                  _rewriter;
     bool                      _gen_pairs_initted;
-    WordGraph<size_t>         _gilman_graph;
+    WordGraph<uint32_t>       _gilman_graph;
     std::vector<std::string>  _gilman_graph_node_labels;
     bool                      _internal_is_same_as_external;
     OverlapMeasure*           _overlap_measure;
@@ -821,7 +821,7 @@ namespace libsemigroups {
     //!
     //! \param
     //! (None)
-    WordGraph<size_t> const& gilman_graph();
+    WordGraph<uint32_t> const& gilman_graph();
 
     //! \brief Return the node labels of the Gilman \ref WordGraph
     //!
@@ -956,7 +956,7 @@ namespace libsemigroups {
   std::ostream& operator<<(std::ostream&,
                            KnuthBendix<Rewriter, ReductionOrder> const&);
 
-  KnuthBendix(congruence_kind)->KnuthBendix<>;
+  KnuthBendix(congruence_kind) -> KnuthBendix<>;
 
   namespace knuth_bendix {
 
@@ -1043,7 +1043,7 @@ namespace libsemigroups {
     template <typename Rewriter, typename ReductionOrder>
     [[nodiscard]] inline auto
     normal_forms(KnuthBendix<Rewriter, ReductionOrder>& kb) {
-      using rx::      operator|;
+      using rx::operator|;
       ReversiblePaths paths(kb.gilman_graph());
       paths.from(0).reverse(kb.kind() == congruence_kind::left);
       if (!kb.presentation().contains_empty_word()) {
